refactor(dsv4): drop int64 cast from flat BF16 scatter
npu_scatter_nd_update_ already accepts int32 indices and non-contiguous
tensors, matching the SFA path. Remove the extra Cast/Copy so ACLGraph
sees the same int32 [T, 1] slots as A5 FP8 / SFA.

+++ tests/ut/attention/test_dsa_attn_kv_plan.py
@@ -144,7 +144,7 @@ def test_bf16_scatter_is_aclgraph_static_and_keeps_pad_slots():
         assert captured["updates"].shape[0] == slot_mapping.shape[0]
         torch.testing.assert_close(
             captured["indices"],
-            torch.tensor([[1], [-1], [-1]], dtype=torch.int64),
+            torch.tensor([[1], [-1], [-1]], dtype=torch.int32),
         )
         torch.testing.assert_close(cache[0, 1], updates[0])
         torch.testing.assert_close(cache[0, 0], slot_zero)
